Share a single Spring context across configprops tests
Collapse the configprops integration tests onto one cached Spring
ApplicationContext. The Converter test and both Service scenarios
previously forked three distinct contexts (differing properties,
imports, and SmartSanitizingFunction beans).

Introduce AbstractConfigPropsIntegrationTest (one @SpringBootTest +
shared @TestPropertySource + @Import) and ConfigPropsTestSupportConfiguration
(shared beans + a single SharedAxelixConfigurationProperties record). The two
sanitization behaviors are now two named DefaultConfigurationPropertiesService
beans (sanitizeAllConfigurationPropertiesService /
explicitlySanitizedConfigurationPropertiesService) in that shared context; the
Service test selects one per @Nested scenario via @Qualifier.

AxelixConfigurationPropertiesEndpointTest keeps its own RANDOM_PORT context;
SmartSanitizingFunctionTest is a plain unit test and is untouched.

spring-test-profiler: configprops contexts 4 -> 2 (one shared + the
isolated EndpointTest).
