+++ sbs/axelix-spring-boot-3-starter/src/test/java/com/axelixlabs/axelix/sbs/spring/core/configprops/DefaultConfigurationPropertiesConverterTest.java
@@ -17,23 +17,15 @@
  */
 package com.axelixlabs.axelix.sbs.spring.core.configprops;
 
-import java.util.List;
-import java.util.Map;
-
 import org.junit.jupiter.api.Test;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.actuate.context.properties.ConfigurationPropertiesReportEndpoint;
 import org.springframework.boot.actuate.context.properties.ConfigurationPropertiesReportEndpoint.ConfigurationPropertiesDescriptor;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.test.context.TestPropertySource;
 
 import com.axelixlabs.axelix.common.api.ConfigurationPropertiesFeed;
 import com.axelixlabs.axelix.common.api.KeyValue;
+import com.axelixlabs.axelix.sbs.spring.core.configprops.ConfigPropsTestSupportConfiguration.SharedAxelixConfigurationProperties;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -42,33 +34,16 @@ import static org.assertj.core.api.Assertions.assertThat;
  *
  * @author Sergey Cherkasov
  */
-@SpringBootTest(properties = "management.endpoint.configprops.show-values=always")
-@TestPropertySource(
-        properties = {
-            "axelix.prop.test.tags.environment=test",
-            "axelix.prop.test.tags.version=1.0.0",
-            "axelix.prop.test.enabled-contexts=user-service,payment-service",
-            "axelix.prop.test.http-client.requests[0].name=user-api",
-            "axelix.prop.test.http-client.requests[0].base-url=https://api.users.example.com/v1",
-            "axelix.prop.test.http-client.requests[0].methods[0].type=GET",
-            "axelix.prop.test.http-client.requests[0].methods[0].retries[0].count=3",
-            "axelix.prop.test.http-client.requests[0].methods[0].retries[0].parameters.timeout=5000",
-            "axelix.prop.test.http-client.requests[0].methods[1].type=POST"
-        })
-@EnableConfigurationProperties(DefaultConfigurationPropertiesConverterTest.AxelixConfigurationProperties.class)
-public class DefaultConfigurationPropertiesConverterTest {
+public class DefaultConfigurationPropertiesConverterTest extends AbstractConfigPropsIntegrationTest {
 
     @Autowired
     private ConfigurationPropertiesReportEndpoint endpoint;
-
-    @Autowired
-    private ConfigurationPropertiesConverter enricher;
 
     @Test
     void shouldReturnConfigurationPropertiesFeed() {
         ConfigurationPropertiesDescriptor defaultDescriptor = endpoint.configurationProperties();
 
-        ConfigurationPropertiesFeed axelixConfPropDescriptor = enricher.convert(defaultDescriptor);
+        ConfigurationPropertiesFeed axelixConfPropDescriptor = converter.convert(defaultDescriptor);
 
         assertThat(axelixConfPropDescriptor).isNotNull();
 
@@ -80,7 +55,7 @@ public class DefaultConfigurationPropertiesConverterTest {
                 .satisfies(bean -> {
 
                     // Bean Name
-                    assertThat(bean.getBeanName()).isEqualTo(AxelixConfigurationProperties.class.getName());
+                    assertThat(bean.getBeanName()).isEqualTo(SharedAxelixConfigurationProperties.class.getName());
 
                     // prefix
                     assertThat(bean.getPrefix()).isEqualTo("axelix.prop.test");
@@ -107,41 +82,5 @@ public class DefaultConfigurationPropertiesConverterTest {
                                     .equals("httpClient.requests[0].methods[0].retries[0].parameters.timeout.value"))
                             .anyMatch(p -> p.getKey().equals("httpClient.requests[0].baseUrl.origin"));
                 });
-    }
-
-    @ConfigurationProperties(prefix = "axelix.prop.test")
-    public record AxelixConfigurationProperties(
-            Map<String, String> tags,
-            List<String> enabledContexts,
-            AxelixConfigurationPropertiesEndpointTest.AxelixConfigurationProperties.HttpClient httpClient) {
-
-        public record HttpClient(
-                List<AxelixConfigurationPropertiesEndpointTest.AxelixConfigurationProperties.Request> requests) {}
-
-        public record Request(
-                String name,
-                String baseUrl,
-                List<AxelixConfigurationPropertiesEndpointTest.AxelixConfigurationProperties.Method> methods) {}
-
-        public record Method(
-                String type,
-                List<AxelixConfigurationPropertiesEndpointTest.AxelixConfigurationProperties.Retry> retries) {}
-
-        public record Retry(Integer count, Map<String, Object> parameters) {}
-    }
-
-    @TestConfiguration
-    static class DefaultDefaultConfigurationPropertiesTestConfiguration {
-
-        @Bean
-        public ConfigurationPropertiesFlattener configurationPropertiesFlattener() {
-            return new DefaultConfigurationPropertiesFlattener();
-        }
-
-        @Bean
-        public ConfigurationPropertiesConverter configurationPropertiesConverter(
-                ConfigurationPropertiesFlattener configurationPropertiesFlattener) {
-            return new DefaultConfigurationPropertiesConverter(configurationPropertiesFlattener);
-        }
     }
 }
